add additional in line comments, manager allocate tests
